Full rework to own region implementation and CommandFramework

+++ src/main/java/de/zeanon/testutils/plugin/commands/testutils/testarea/DeleteArea.java
@@ -1,9 +1,8 @@
 package de.zeanon.testutils.plugin.commands.testutils.testarea;
 
-import com.sk89q.worldedit.bukkit.BukkitWorld;
-import com.sk89q.worldedit.world.World;
 import de.zeanon.testutils.TestUtils;
 import de.zeanon.testutils.plugin.utils.GlobalMessageUtils;
+import de.zeanon.testutils.plugin.utils.enums.AreaName;
 import de.zeanon.testutils.plugin.utils.region.DefinedRegion;
 import java.io.File;
 import java.io.IOException;
@@ -18,40 +17,31 @@ import org.jetbrains.annotations.Nullable;
 @UtilityClass
 public class DeleteArea {
 
-	public void execute(final @NotNull String[] args, final @NotNull Player p) {
-		if (args.length == 2) {
-			if (DeleteArea.remove(new BukkitWorld(p.getWorld()), args[1])) {
-				p.sendMessage(GlobalMessageUtils.messageHead
-							  + ChatColor.RED + "You deleted the testarea '" + ChatColor.DARK_RED + args[1] + ChatColor.RED + "'.");
-			} else {
-				p.sendMessage(GlobalMessageUtils.messageHead
-							  + ChatColor.RED + "Testarea '" + ChatColor.DARK_RED + args[1] + ChatColor.RED + "' does not exist and thus could not be deleted.");
-			}
-		} else if (args.length == 1) {
+	public void execute(final @NotNull AreaName name, final @NotNull Player p) {
+		if (DeleteArea.remove(name.getName())) {
 			p.sendMessage(GlobalMessageUtils.messageHead
-						  + ChatColor.RED + "Missing argument for "
-						  + ChatColor.YELLOW + "<"
-						  + ChatColor.GOLD + "name"
-						  + ChatColor.YELLOW + ">");
+						  + ChatColor.RED + "You deleted the testarea '" + ChatColor.DARK_RED + name + ChatColor.RED + "'.");
 		} else {
 			p.sendMessage(GlobalMessageUtils.messageHead
-						  + ChatColor.RED + "Too many arguments.");
+						  + ChatColor.RED + "Testarea '" + ChatColor.DARK_RED + name + ChatColor.RED + "' does not exist and thus could not be deleted.");
 		}
 	}
 
-	private boolean remove(final @NotNull World world, final @NotNull String name) {
+	private boolean remove(final @NotNull String name) {
 		final @Nullable DefinedRegion southRegion = de.zeanon.testutils.plugin.utils.region.RegionManager.getRegion(name + "_south");
 		final @Nullable DefinedRegion northRegion = de.zeanon.testutils.plugin.utils.region.RegionManager.getRegion(name + "_north");
-		if (southRegion != null && northRegion != null) {
-			de.zeanon.testutils.plugin.utils.region.RegionManager.removeRegion(southRegion);
-			de.zeanon.testutils.plugin.utils.region.RegionManager.removeRegion(northRegion);
-
+		if (southRegion != null && northRegion != null && de.zeanon.testutils.plugin.utils.region.RegionManager.removeRegion(southRegion) && de.zeanon.testutils.plugin.utils.region.RegionManager.removeRegion(northRegion)) {
 			try {
-				final @NotNull File resetFolder = new File(TestUtils.getInstance().getDataFolder(), "/TestAreas/" + world.getName() + "/" + name.substring(0, name.length() - 6));
+				final @NotNull File resetFolder = new File(TestUtils.getInstance().getDataFolder(), "/TestAreas/" + name.substring(0, name.length() - 6));
 				if (resetFolder.exists() && resetFolder.isDirectory()) {
 					FileUtils.deleteDirectory(resetFolder);
 				}
-				final @NotNull File backUpFolder = new File(TestUtils.getInstance().getDataFolder(), "/Backups/" + world.getName() + "/" + name.substring(0, name.length() - 6));
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+
+			try {
+				final @NotNull File backUpFolder = new File(TestUtils.getInstance().getDataFolder(), "/Backups/" + name.substring(0, name.length() - 6));
 				if (backUpFolder.exists() && backUpFolder.isDirectory()) {
 					FileUtils.deleteDirectory(backUpFolder);
 				}
